+++ wallet/api/pipeline/sequential.go
@@ -16,9 +16,9 @@ var (
 // SequentialPipeline is a pipeline built to handle one request at a time.
 // Concurrent requests are not supported and will result in errors.
 type SequentialPipeline struct {
-	// clientCtx is the context used to listen to the client-side cancellation
+	// userCtx is the context used to listen to the user-side cancellation
 	// requests. It interrupts the wait on responses.
-	clientCtx context.Context
+	userCtx context.Context
 
 	receptionChan chan<- Envelope
 	responseChan  <-chan Envelope
@@ -65,8 +65,8 @@ func (s *SequentialPipeline) RequestWalletConnectionReview(ctx context.Context, 
 		select {
 		case <-ctx.Done():
 			return false, api.ErrRequestInterrupted
-		case <-s.clientCtx.Done():
-			return false, api.ErrConnectionClosed
+		case <-s.userCtx.Done():
+			return false, api.ErrUserCloseTheConnection
 		case response := <-s.responseChan:
 			if response.TraceID != traceID {
 				return false, ErrTraceIDMismatch
@@ -97,8 +97,8 @@ func (s *SequentialPipeline) RequestWalletSelection(ctx context.Context, traceID
 		select {
 		case <-ctx.Done():
 			return api.SelectedWallet{}, api.ErrRequestInterrupted
-		case <-s.clientCtx.Done():
-			return api.SelectedWallet{}, api.ErrConnectionClosed
+		case <-s.userCtx.Done():
+			return api.SelectedWallet{}, api.ErrUserCloseTheConnection
 		case response := <-s.responseChan:
 			if response.TraceID != traceID {
 				return api.SelectedWallet{}, ErrTraceIDMismatch
@@ -128,8 +128,8 @@ func (s *SequentialPipeline) RequestPassphrase(ctx context.Context, traceID, wal
 		select {
 		case <-ctx.Done():
 			return "", api.ErrRequestInterrupted
-		case <-s.clientCtx.Done():
-			return "", api.ErrConnectionClosed
+		case <-s.userCtx.Done():
+			return "", api.ErrUserCloseTheConnection
 		case response := <-s.responseChan:
 			if response.TraceID != traceID {
 				return "", ErrTraceIDMismatch
@@ -161,8 +161,8 @@ func (s *SequentialPipeline) RequestPermissionsReview(ctx context.Context, trace
 		select {
 		case <-ctx.Done():
 			return false, api.ErrRequestInterrupted
-		case <-s.clientCtx.Done():
-			return false, api.ErrConnectionClosed
+		case <-s.userCtx.Done():
+			return false, api.ErrUserCloseTheConnection
 		case response := <-s.responseChan:
 			if response.TraceID != traceID {
 				return false, ErrTraceIDMismatch
@@ -196,8 +196,8 @@ func (s *SequentialPipeline) RequestTransactionSendingReview(ctx context.Context
 		select {
 		case <-ctx.Done():
 			return false, api.ErrRequestInterrupted
-		case <-s.clientCtx.Done():
-			return false, api.ErrConnectionClosed
+		case <-s.userCtx.Done():
+			return false, api.ErrUserCloseTheConnection
 		case response := <-s.responseChan:
 			if response.TraceID != traceID {
 				return false, ErrTraceIDMismatch
@@ -231,8 +231,8 @@ func (s *SequentialPipeline) RequestTransactionSigningReview(ctx context.Context
 		select {
 		case <-ctx.Done():
 			return false, api.ErrRequestInterrupted
-		case <-s.clientCtx.Done():
-			return false, api.ErrConnectionClosed
+		case <-s.userCtx.Done():
+			return false, api.ErrUserCloseTheConnection
 		case response := <-s.responseChan:
 			if response.TraceID != traceID {
 				return false, ErrTraceIDMismatch
@@ -262,9 +262,9 @@ func (s *SequentialPipeline) NotifyTransactionStatus(ctx context.Context, traceI
 	}
 }
 
-func NewSequentialPipeline(clientCtx context.Context, receptionChan chan<- Envelope, responseChan <-chan Envelope) *SequentialPipeline {
+func NewSequentialPipeline(userCtx context.Context, receptionChan chan<- Envelope, responseChan <-chan Envelope) *SequentialPipeline {
 	return &SequentialPipeline{
-		clientCtx:     clientCtx,
+		userCtx:       userCtx,
 		receptionChan: receptionChan,
 		responseChan:  responseChan,
 	}